Changes to plugins.user and plugins.password
The class plugins.user.User was changed so the name and age are no
longer required. This also lead to a minor update of challenge 001e
and challenge 005e.
The class plugins.user.UserDatabase was then also adjusted so the
only required User attributes (by default) are the username and
the password. A method was added to change the required User
attributes of users in a UserDatabase.
All plugins.user unit tests were updated accordingly.

The plugin plugins.authorization was renamed to plugins.password,
and its unittests were updated accordingly.