Removed unnecessary plugins
I was starting to put everything in plugins. That is not the idea of a
plugin, a plugin should only be created if part of a solution is used
in multiple challenge solutions. Therefore, plugins that contained
functions or classes that are only used in one of the currently solved
challenges were removed, and their contents were moved to the
respective challenge module.