Changed definition of 'alphanumeric' in challenge 019e
The built-in string method 'str.isalnum()' checks if a character is
alphanumeric. The challenge solution was edited to incorporate this
method. But since the method considers letters with accents to be
alphanumeric, the definition of 'alphanumeric' had to be changed in
the challenge solution, and the results are slightly different.