Added method for automatically preparing a challenge solution module blueprint
The main dailyprogrammer execution script was updated to include a
way for automatically generating a challenge solution module blueprint.

A command-line argument 'prepare' was added which, combined with a
challenge ID, automatically finds the url to the correct submission
on Reddit using BeatifulSoup and then extracts the challenge's
information from the submission using PRAW. Finally, this information
is used to format a challenge template file.

Some safeties have been built in. Preparing a blueprint for a challenge
is not possible if there is already a module with that challenge ID.
If multiple challenges with the ID are found, a warning is printed and
the first one is automatically picked for generation. The others will
have to be created manually, with a non-default challenge ID.

This change also meant that an update was required for how challenge
solutions are executed. For this, a command-line argument 'execute'
was added which, combined with a challenge ID, will execute that
challenge (if it exists).