Moved TextTriangle to plugin 'asciiart' in preparation of challenge 035e
The class TextTriangle, originally created in challenge 017e, was moved
to a new plugin called 'asciiart'. This was deemed appropriate because it
could easily be used as a base for solving challenge 035e, while still
maintaining perfect compatibility with challenge 017e. The required
modifications were made and unit tests were created and successfully
executed. Documentation will be updated in next commit, together with
the submission of the solution to challenge 035e.