Added challenge execution tests
Sometimes a challenge motivates me to update one of the plugins. It's possible,
however, that such an update breaks execution of older challenges that use the
affected plugin. In order to check whether or not this is the case, I had to
execute each challenge individually and see if everything still works properly.
As the amount of solved challenges increases, this will get more and more
annoying to do.

Because of this, I decided to add automatic execution tests to the testing
framework. In these tests, each challenge's example run (see documentation) is
simulated, thus testing successful execution and compatibility with the latest
versions of the plugins. Correctness of the output is, in most cases, *not*
(yet) tested. Although for now I deem the likelihood of a successful but
incorrect execution quite low.